requirements.txt
opencv-contrib-python
pillow
pytesseract
playsound
pyautogui
opencv-python